refactor and rename clicked user profile functions
